+++ src/components/locations/Brewery.js
@@ -1,6 +1,6 @@
 import { useEffect, useState } from "react";
 import { useSelector } from "react-redux";
-import LocationMarker from "../map/LocationMarker"
+import BreweryMarker from "./BreweryMarker"
 
 const Brewery = ({brewery}) => {
   const {lat, lng} = brewery.locationProperties;
@@ -13,8 +13,10 @@ const Brewery = ({brewery}) => {
     }
   }, [breweryStore, brewery]);
 
+  
+
   return(<>
-    {brewery && brewery.locationProperties && <LocationMarker markerPosition={{lat, lng}} brewery={brewery} closest={isClosest}/>}
+    {brewery && brewery.locationProperties && <BreweryMarker markerPosition={{lat, lng}} brewery={brewery} closest={isClosest}/>}
     </>
   )
 }
