+++ src/components/locations/LocationMarker.js
@@ -1,10 +1,9 @@
-import { useCallback, useEffect, useRef, useState } from 'react';
+import { useEffect, useRef, useState } from 'react';
 import { useMapEvents } from 'react-leaflet';
 import { Marker, Popup } from 'react-leaflet';
 import L from 'leaflet';
 
 import pointingDown from '../../assets/you-are-here-marker.svg';
-import BreweryPopup from '../locations/BreweryPopup';
 
 L.Icon.Default.imagePath='img/'
 
@@ -20,7 +19,7 @@ const pointingDownIcon = new L.Icon({
     className: 'leaflet-div-icon'
 });
 
-function LocationMarker({ markerPosition, brewery, closest }) {
+function LocationMarker({ markerPosition }) {
   const markerRef = useRef(null);
   const [position, setPosition] = useState(null);
   
@@ -34,33 +33,18 @@ function LocationMarker({ markerPosition, brewery, closest }) {
     },
   })
 
-  function openPopup(closest) {
-    if(closest) {
-      markerRef.current.openPopup();
-    }
-  }
-
-  useCallback(
-    openPopup(closest)
-  , [closest]);
-
   useEffect(() => {
     //watch position change of marker
     setPosition([markerPosition.lat, markerPosition.lng])
   }, [markerPosition])
 
-  return position === null ? null : (<>
-    {brewery ? 
-      <Marker ref={markerRef} position={position} popupOpen={closest}>
-        <BreweryPopup brewery={brewery}/>
-      </Marker> :
+  return position === null ? null : (
       <Marker ref={markerRef} position={position} icon={pointingDownIcon}>
         <Popup>
           You are here!
         </Popup>
       </Marker>
-    }
-  </>)
+  )
 }
 
 export default LocationMarker;
